refactor(settings): extract lib/settings.php + ADR-001 type dispatch (sub of #907) — closes #915
Moves the 14 settings functions out of lib.php into a new lib/settings.php
module (ADR-004 Phase 5 Task 5.2b). Functions stay in the global namespace
per ADR-004 Option E; ipam_config_stale_keys() deliberately stays in lib.php
(a config concern, not a setting).

ADR-001 — the registry becomes DB-backed. ipam_setting_definitions() now
reads the setting_definitions table and reconstructs the v3.29.0 array shape,
adding a NEW logical_type key (the 11-value type) while keeping type as the
4-value STORAGE type so every existing caller and the settings.type CHECK
keep working. The frozen v3.29.0 PHP registry is preserved as
ipam_setting_definitions_seed() — the install-time seed source and the
fallback used when the table is absent (fresh install pre-migration) or
empty. The 3.30.0-setting-definitions migration closure now calls
ipam_setting_definitions_seed() to break the circular bootstrap dependency.

New 11-type dispatch helpers: ipam_setting_storage_type() (logical -> 4-value
storage type) and ipam_setting_validate() (per-type validation for int/bool/
json/enum/secret/string/url/email/timezone/cidr/datetime).

#915 — the 5-param, 4-mode ipam_setting_cache_storage() is replaced by the
focused trio ipam_setting_cache_get / _set / _clear over one shared backing
store. ipam_setting_cache_bust() stays as the public name and delegates to
_clear(); _clear() also resets the definitions DB-read memo.

ADR-003 — $GLOBALS['config'] reads in ipam_setting() and
ipam_setting_deprecated_keys() route through ipam_config(). $GLOBALS['db']
is deliberately retained (runtime PDO handle, not config).

Tests: tests/SettingDispatchTest.php (all 11 logical types — storage mapping,
round-trip, validation, enum options) written TDD-first; tests/Unit/
SettingsExtractionParityTest.php mirrors the presentation/audit parity tests.
Callers of the removed ipam_setting_cache_storage() updated across 6 test
files. Full local gate green: 1307 PHPUnit tests, PHPStan L9, phpcs,
lib-module-linter, migration-linter; bootstrap-app sqlite confirmed the
seeded DB-backed ipam_setting_definitions() returns all 11 logical types
with the storage type preserved (103 entries equivalent to the seed).
