+++ tests/NotificationDispatcherTest.php
@@ -197,7 +197,7 @@ class NotificationDispatcherTest extends TestCase
     protected function tearDown(): void
     {
         unset($GLOBALS['db'], $GLOBALS['config']);
-        ipam_setting_cache_storage('__CLEAR__', true);
+        ipam_setting_cache_bust();
     }
 
     // ---------------------------------------------------------------------
